# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Participación Ciudadana Documentación API",
        "description": "Endpoints de la app Participación Ciudadana",
        "contact": {
            "email": "ciudadana@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001/api/v1",
            "description": "Endpoints de la app Participación Ciudadana"
        }
    ],
    "paths": {
...
        "/password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Forgot the password",
                "description": "If user forgot his password, a new password will be generated and send by mail",
                "operationId": "forgotPassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email of user",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Your password has been changed, please check your email"
                    },
                    "404": {
                        "description": "User's email not provided"
                    }
                }
            }
        },
...
    },
...
}